pyproject.toml
[project]
name = "python-example"
version = "0.1.0"
description = "Digital photo frame (carousel) with Google Drive integration."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "google-api-python-client",
    "google-auth",
    "Pillow",
    "PyQt5"
]
